Fix followers show, Add polymorphic assoc 4 shouts
Lots of fun making the database migration to preserve user data while
allowing the existing shouts table to be reformatted for the polymorphic
association.

This broke the ability to create and display a user's shouts, so
recreated that.

Also had missed the ".all" on the
@followers = current_user.followers.all
in FollowersController. Adding that in enabled the iterator on the show
page to do its thing.